pyproject.toml
[project]
name = "timestep-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.2",
    "openai-agents[sqlalchemy]>=0.5.0",
    "openai-chatkit>=1.1.2",
    "supabase>=2.22.4",
    "asyncpg>=0.29.0",
    "ollama>=0.6.0",
]

[tool.uv.sources]
openai-agents = { git = "https://github.com/mjschock/openai-agents-python.git" }
